third_party/WebKit/LayoutTests/dom/document_fragment/get-element-by-id-expected.txt

Tests that getElementById() API is exposed on DocumentFragment nodes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS fragment.getElementById('divID') is div
PASS fragment.getElementById('aID') is a
PASS fragment.getElementById('notInFragment') is null
PASS fragment.getElementById('doesNotExist') is null
PASS fragment.getElementById('duplicateId1') is span
PASS fragment.getElementById('duplicateId2') is h1
PASS successfullyParsed is true

TEST COMPLETE
